bring the queue into its own files
